setup runtime for rococo
