+++ artifacts/bitgreen-rococo-genesis
@@ -1,0 +1,1 @@
+0x0000000000000000000000000000000000000000000000000000000000000000009165ed6cc2baaa2ea7cdd560309f1baafa323d8500f455b6c1d1258ac60009c303170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c11131400
