--- polkadot-launch/config.json
@@ -1,49 +1,0 @@
-{
-	"relaychain": {
-		"bin": "../../../polkadot/target/release/polkadot",
-		"chain": "rococo-local",
-		"nodes": [
-			{
-				"name": "alice",
-				"wsPort": 9944,
-				"port": 30444
-			},
-			{
-				"name": "bob",
-				"wsPort": 9955,
-				"port": 30555
-			}
-		],
-		"genesis": {
-			"runtime": {
-				"runtime_genesis_config": {
-					"configuration": {
-						"config": {
-							"validation_upgrade_frequency": 10,
-							"validation_upgrade_delay": 10
-						}
-					}
-				}
-			}
-		}
-	},
-	"parachains": [
-		{
-			"bin": "../target/release/bitgreen-parachain",
-			"chain" : "dev",
-			"id": "2000",
-			"balance": "1000000000000000000000",
-			"nodes": [
-				{
-					"wsPort": 9988,
-					"port": 31200,
-					"name": "alice",
-					"flags": ["--", "--execution=wasm"]
-				}
-			]
-		}
-	],
-	"types": {},
-	"finalization": false,
-	"simpleParachains": []
-}
