+++ runtime/rococo/Cargo.toml
@@ -54,6 +54,7 @@ sp-session = { git = "https://github.com/paritytech/substrate", default-features
 sp-std = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.29" }
 sp-transaction-pool = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.29" }
 sp-version = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.29" }
+pallet-treasury = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.29" }
 
 # Polkadot
 pallet-xcm = { git = "https://github.com/paritytech/polkadot", default-features = false, branch = "release-v0.9.29" }
@@ -90,6 +91,7 @@ pallet-carbon-credits-pool = { default-features = false, version = '0.0.1', path
 pallet-transaction-pause = { default-features = false, version = '0.0.1', path = "../../pallets/transaction-pause" }
 pallet-vesting-contract = { default-features = false, version = '0.0.1', path = "../../pallets/vesting-contract" }
 primitives = { package = "bitgreen-primitives", path = "../../primitives", default-features = false }
+pallet-parachain-staking = { default-features = false, version = '0.0.1', path = "../../pallets/parachain-staking" }
 
 [features]
 default = [
@@ -144,6 +146,7 @@ std = [
   "xcm/std",
   "pallet-uniques/std",
   "pallet-membership/std",
+  "pallet-treasury/std",
 
   "orml-traits/std",
   "orml-tokens/std",
@@ -157,6 +160,7 @@ std = [
   "pallet-carbon-credits-pool/std",
   "pallet-transaction-pause/std",
   "pallet-vesting-contract/std",
+  "pallet-parachain-staking/std"
 ]
 
 runtime-benchmarks = [
